requirements.txt
functions-framework==3.*
google-cloud-firestore==2.10.0
pandas==1.5.3